requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
pandas==2.1.3
termcolor==2.3.0
numpy==1.24.3
pandas-ta==0.3.14b0
schedule==1.2.1
python-dateutil==2.8.2
backoff==2.2.1
youtube-transcript-api==0.6.1
openai==1.3.5
# Add any other dependencies your agents need
